fix: reorganize & rename ZapData into Context
