+++ src/providers/ZapInitProvider/ModularZaps/base.ts
@@ -1,6 +1,8 @@
 import { Instruction, MultichainSmartAccount } from '@biconomy/abstractjs';
 import { Route } from '@lifi/sdk';
+import { EVMAddress } from 'src/types/internal';
 import { ProjectData } from 'src/types/questDetails';
+import { zeroAddress } from 'viem';
 import { AbiEntry, ZapDataResponse } from './zap.jumper-backend';
 
 export interface SendCallsExtraParams {
@@ -44,5 +46,87 @@ export interface ZapDefinition {
 
 export type ZapInstruction = (
   oNexus: MultichainSmartAccount,
-  params: ZapExecutionContext,
+  context: ZapExecutionContext,
 ) => Promise<Instruction[] | null>;
+
+export const makeZapExecutionContext = (
+  params: ValidatedSendCallsExtraParams,
+): ZapExecutionContext => {
+  const market = params.zapData.market;
+  if (!market) {
+    throw new Error('Market not found in zap data');
+  }
+
+  const getAbiAddress = (fct: AbiEntry) => {
+    if (fct.contract) {
+      const contracts = market.contracts;
+      if (!contracts) {
+        throw new Error('Contracts not found in market');
+      }
+
+      const v = contracts[fct.contract];
+      if (!v) {
+        throw new Error(`Contract ${fct.contract} not found in market`);
+      }
+      return v;
+    }
+    return market.address;
+  };
+
+  const getDepositAddress = () => {
+    return getAbiAddress(params.zapData.abi.deposit);
+  };
+
+  return {
+    ...params,
+    getAbiAddress,
+    getDepositAddress,
+  };
+};
+
+export const isValidParams = (
+  sendCallsExtraParams: SendCallsExtraParams,
+): sendCallsExtraParams is ValidatedSendCallsExtraParams => {
+  if (!sendCallsExtraParams.currentRoute) {
+    throw new Error('Current route is not set');
+  }
+
+  const {
+    chainId: currentChainId,
+    address: currentAddress,
+    zapData: integrationData,
+    projectData,
+  } = sendCallsExtraParams;
+
+  if (!currentChainId || !currentAddress) {
+    throw new Error('Missing chainId or address');
+  }
+
+  const currentRouteFromToken = sendCallsExtraParams.currentRoute.fromToken;
+  const depositAddress = integrationData.market?.address as EVMAddress;
+  const depositToken = integrationData.market?.depositToken?.address;
+  const depositTokenDecimals = integrationData.market?.depositToken.decimals;
+  const depositChainId = projectData.chainId;
+
+  if (!depositChainId) {
+    throw new Error('Deposit chain id is undefined.');
+  }
+
+  if (!depositAddress || !depositToken) {
+    throw new Error('Deposit address or token is undefined.');
+  }
+
+  if (!depositTokenDecimals) {
+    throw new Error('Deposit token decimals is undefined.');
+  }
+
+  // @Note this works only for EVM chains
+  const isNativeSourceToken = currentRouteFromToken.address === zeroAddress;
+  console.warn('Using native source token:', isNativeSourceToken);
+
+  if (isNativeSourceToken) {
+    throw new Error('Native source token is not supported.');
+  }
+
+  return true;
+};
